Refactoring to new structure is operational.
